moved storeProvider under components
