+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import { Geist, Geist_Mono } from "next/font/google";
 import { ToastContainer } from "react-toastify";
 
 import "./globals.css";
-import StoreProvider from "./StoreProvider";
+import StoreProvider from "../components/storeProvider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
