+++ src/components/storeProvider.tsx
@@ -1,15 +1,11 @@
 "use client";
 
-import { useRef } from "react";
+import { ReactNode, useRef } from "react";
 import { Provider } from "react-redux";
 
 import { makeStore, AppStore } from "../lib/store";
 
-export default function StoreProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function StoreProvider({ children }: { children: ReactNode }) {
   const storeRef = useRef<AppStore>(undefined);
   if (!storeRef.current) {
     // Create the store instance the first time this renders
